refactor: Address code review feedback for order sync implementation
Based on review comments, implemented the following improvements:

1. Code Quality and Maintainability:
   - Refactored 572-line sync_router.py into organized sub-routers
   - Split into status, config, conflict, and main routers for better separation
   - Improved error handling with specific exception types (HTTPError, TimeoutError, etc.)
   - Extracted magic constants to configurable settings in config.py

2. Testing:
   - Added comprehensive integration tests for conflict resolution
   - Added scheduler lifecycle tests (startup/shutdown)
   - Added tests for manual sync triggers and error scenarios

3. Database/Migrations:
   - Fixed Alembic migration down_revision to properly chain
   - Moved migration to correct location with proper naming

4. Performance & Scaling:
   - Made concurrency limits configurable (SYNC_CONCURRENT_ORDERS)
   - All intervals and timeouts now environment-configurable
   - Added configuration for retry backoff, health check intervals, etc.

5. Frontend:
   - Made dashboard polling interval configurable via environment variable

This refactoring improves maintainability, testability, and production-readiness
while addressing all high-priority review concerns.
