+++ backend/modules/orders/routers/sync/__init__.py
@@ -1,0 +1,14 @@
+# backend/modules/orders/routers/sync/__init__.py
+
+"""
+Order synchronization routers module.
+
+Provides organized sub-routers for different sync functionalities.
+"""
+
+from .status_router import router as status_router
+from .config_router import router as config_router
+from .conflict_router import router as conflict_router
+from .main_router import router as sync_router
+
+__all__ = ['sync_router', 'status_router', 'config_router', 'conflict_router']
